refactor: streamline user data fetching and SEO meta management in app, index, and profile pages
